add feature flag for configure public url
