+++ src/locales/en-US/admin.ftl
@@ -1082,6 +1082,9 @@ moderate-user-drawer-created-at =
   .title = Account creation date
 moderate-user-drawer-member-id =
   .title = Member ID
+moderate-user-drawer-external-profile-URL =
+  .title = External profile URL
+moderate-user-drawer-external-profile-URL-link = External profile URL
 moderate-user-drawer-tab-all-comments = All Comments
 moderate-user-drawer-tab-rejected-comments = Rejected
 moderate-user-drawer-tab-account-history = Account History
